refactor: alimento search and display logic
Moved and improved the alimento search screen, consolidating logic from montar_prato/buscar_alimentos_screen.dart into buscar/buscar_alimento_screen.dart. Added AlimentoDisplayHelper for friendly formatting and image URL handling. Updated API base URL usage, caching of alimento metadata, and refactored repository to preserve and restore more alimento fields. Removed unused buscar_alimentos_screen.dart and updated router to use the new screen.
